requirements.txt
annotated-types==0.7.0
autocuda==0.16
blis==1.1.0
boostaug==2.3.5
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
distlib==0.3.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
filelock==3.15.1
findfile==2.0.1
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
metric_visualizer==0.9.14.post1
mpmath==1.3.0
murmurhash==1.0.11
natsort==8.4.0
networkx==3.4.2
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.2.2
praw==7.8.1
prawcore==2.4.0
preshed==3.0.9
protobuf==3.20.3
pyabsa==2.4.1.post1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-warmup==0.2.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.1
scikit-learn==1.6.0
scipy==1.15.0
sentencepiece==0.2.0
seqeval==1.2.2
setuptools==75.7.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
sympy==1.13.3
tabulate==0.9.0
termcolor==2.5.0
thinc==8.3.3
threadpoolctl==3.5.0
tikzplotlib==0.10.1
tokenizers==0.21.0
torch==2.2.2
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.3.0
virtualenv==20.26.2
wasabi==1.1.3
weasel==0.4.1
webcolors==24.11.1
websocket-client==1.8.0
wrapt==1.17.0
XlsxWriter==3.2.0
